examples/semifairr_project/pyproject.toml
[project]
name = "climsim-analysis"
version = "0.3.0"
description = "Climate simulation post-processing pipeline for ClimSim benchmark data"
authors = [
    {name = "C. de Wit", email = "c.dewit@example.nl"},
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "xarray>=2024.1",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "netcdf4>=1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
